Alyssa's Update 3.0
Improved the review section of an event by adding star rating and users are able to reply to another user's review. Users can see other users' profile pictures in the review section (added a placeholder image if user does not have a profile pic or is not logged in). Also, when buying tickets to an event, users can pick which type of ticket they want for the event by clicking on the drop down.